+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 0.0.4
+
+* Upgrade of Drawer Manager (0.0.4)
+* Breaking change: Adding IntegrationTestHelperBinding
+* Adding BaseIntegrationTest.takeScreenshot(String)
+
+## 0.0.3
+
+* Updated example file paths
+* Added scrollToListTypeItemText
+* Updated README / SUPPORT
+
+## 0.0.2
+
+* Updating README and SUPPORT
+
 ## 0.0.1
 
 * Initial release of the Integration Test Helper package
+